feat: hydration with zoneless. Having to use HTTP Client
follow up PR to make the httpclient eden type safe
